+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Git clone with Jujutsu: Clone Git repositories directly from File > New > Project from Version Control using `jj git clone`
+- Initial jj setup: Prompt to configure user.name and user.email when jj is detected but not configured, pre-populating from git global config
 
 ## [0.5.9] - 2026-03-29